+++ tests/plugins.test.ts
@@ -1,48 +1,7 @@
 import { assertEquals as equals } from "../deps/assert.ts";
-import * as plugins from "../plugins.ts";
 import { pluginNames } from "../cli/utils.ts";
 
-const allNames = Object.keys(plugins);
 const totalPlugins = Array.from(Deno.readDirSync("plugins")).length;
-
-Deno.test("Plugins module", () => {
-  equals(allNames.length, totalPlugins);
-
-  equals(allNames, [
-    "attributes",
-    "basePath",
-    "codeHighlight",
-    "date",
-    "esbuild",
-    "eta",
-    "imagick",
-    "inline",
-    "json",
-    "jsx",
-    "liquid",
-    "markdown",
-    "metas",
-    "modifyUrls",
-    "modules",
-    "netlifyCMS",
-    "nunjucks",
-    "onDemand",
-    "paginate",
-    "parcelCss",
-    "postcss",
-    "prism",
-    "pug",
-    "relativeUrls",
-    "resolveUrls",
-    "sass",
-    "search",
-    "slugifyUrls",
-    "svgo",
-    "terser",
-    "url",
-    "yaml",
-  ]);
-});
 
 Deno.test("Plugins list in init", () => {
   equals(pluginNames.length, totalPlugins - 8);
@@ -60,6 +19,7 @@ Deno.test("Plugins list in init", () => {
     "liquid",
     "metas",
     "modify_urls",
+    "multilanguage",
     "netlify_cms",
     "on_demand",
     "parcel_css",